Standard Readme file
The goal of this README file is to be used as a template for new projects and inspiration for existing README files. 

The intention of the README is not to act as a hard guideline on what should or should not go into a README. Instead, it offers a space to include the most relevant information. It is the responsibility of the developers of each project to decide what is the most pertinent information and modify this template to accommodate the needs.