Documentaton Rails deployment environemnents.
Initial draft documentaton about the deployment environemnents for Rails
apps.

The goal is for this document to provide good guidlenes for the
standard services and configurations that we want to have in place for
our production Rails applications.

This is very much a work in progress and contains several questions that
need to be answered.

Please provide answers and suggestions.