Improve L4 - L6
We've been working on the career ladder for well over a year now. This means that we have gathered a ton of feedback.
We need more challenging levels if we want the career ladder to be useful. It needs to represent a real progression on someone's career where they can bring the maximum value.

This modification aims to give a better overview of what brings value to Abtion. It should be a base for a conversation with your direct manager on how to progress in your career.