Update Software Engineer Progression Framework
Hey users!

We’re here with a big update with ~80% breaking changes.

But why?
Back when we started implementing eXtreme Programming we launched the career ladder as a tool for Abtioneers to see what was expected of them.
With time we’ve learned what works for us and the career ladder should reflect that.

As with everything else in Abtion, we’ve received feedback and ideas for improving our EM toolset and some of the recurring subjects were:
- It was hard to figure out how to progress and why the progression didn’t happen in the expected tempo.
- The career ladder was too specific to a certain T-shirt shaped profile.
- Some of the expectations weren’t realistic in our current setup.

Based on this feedback, we’ve brought the career ladder - now called “The Progression Framework” - up-to-date.

Changelog:
- Researched our projects’ lifecycle and clarified each role’s responsibilities in that lifecycle.
- Updated the framework to reflect this.
- Renamed and indexed roles.
- Simplified structure by dividing it into digestible chapters.
- Fixed a bug where people felt stuck in a role.
- Added “Additional reading”-section containing useful subjects for each role.
- Added emojis to all chapter names because they’re fun 💃
- Removed emojis again because it was hard to decide on the right ones [REDACTED]